Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > It's not clear to me why we need this?

We have custom widget, based on Composite. This custom widget uses remote API (not Composite LCA) and every property, which we want to sync with the client must be rendered with the remote API. The supper call only changes the server-side property, but does not render anything.

Patch-set: 1